progress: split progres.go into {progress,run}.go
This is preparing to extract the run part of the progress package
into its own code. Currently there is an unfortunate mix of
concerns here, i.e. `progress` should be generic and does not
need to know details about osbuild and it really should be
its own package. Until that is done this split already helps
(a bit).
